% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Get variable importance.}
\usage{
importance(x, ...)
}
\arguments{
\item{x}{A \code{simplerandomforest} object.}

\item{...}{Further arguments passed down to and from other methods.}
}
\description{
Get variable importance.
}
