% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Predict on new data with a simplerandomforest model.}
\usage{
predict(model, x, num.threads = NULL, ...)
}
\arguments{
\item{model}{A \code{simplerandomforest} object.}

\item{x}{Data for prediction.}

\item{num.threads}{Number of threads to use. Default is number of cores available.}

\item{...}{Further arguments passed to and from other methods.}
}
\description{
Predict on new data with a simplerandomforest model.
}
