% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplerandomforest.R, R/zzz.R
\docType{package}
\name{simplerandomforest}
\alias{simplerandomforest}
\alias{simplerandomforest}
\alias{simplerandomforest-package}
\title{A basic Random Forest implementation.}
\usage{
simplerandomforest(x, y, num.trees = 500, mtry = NULL, replace = FALSE,
  sample.fraction = ifelse(replace, 1, 0.632),
  permutation.importance = FALSE, num.threads = NULL)
}
\arguments{
\item{x}{Training data.}

\item{y}{Reponse variables.}

\item{num.trees}{Number of decision trees to train.}

\item{mtry}{Number of variable to consider at each split. Default is square root of number of independent variables.}

\item{replace}{Sample with replacement.}

\item{sample.fraction}{Fraction of observations to sample.}

\item{permutation.importance}{Compute permutation importance.}

\item{num.threads}{Number of threads to use. Default is number of cores available.}
}
\description{
A basic Random Forest implementation.

simplerandomforest
}
\author{
Simon J. Larsen <simonhffh@gmail.com>
}
